Feat: Refactor get post item use case
